Big chnages
Added support for docker and or kubernetes - if a environment variable 'K8S' is set the app will listen ip 0.0.0.0 (any ip) and default http port 80 and will expect configuration file /etc/quiz_engin_config.ini.

The file size is checked and if it is bigger then 25MB a warning is generated. The size can be changed.

Added button to export the quiz in the editor

Small changes on the css/js/html nothing too noticeable.