future proofing added new category of questions
Added the theoretical idea to add new type of questions.
Possibility to make the building and checking functions in quiz.js  much simpler
Same apply to the template question_editor.tpl

SQL DB schema is not compatible with older version since there is a new column.

When A new type of question is added the compatibility with old txt based exams will de dropped. But since this types of exams are read only, any change equals to conversion to the new DB, this should not be a big problem. Maybe existing exams in json/pkl will need additional work.